+++ src/main/java/com/barion/dungeons_enhanced/world/structures/prefabs/DEUnderwaterStructure.java
@@ -1,9 +1,9 @@
 package com.barion.dungeons_enhanced.world.structures.prefabs;
 
 import com.barion.dungeons_enhanced.DEStructures;
-import com.barion.dungeons_enhanced.DEUtil;
 import com.barion.dungeons_enhanced.world.gen.DETerrainAnalyzer;
 import com.barion.dungeons_enhanced.world.structures.prefabs.utils.DEStructurePiece;
+import com.barion.dungeons_enhanced.world.structures.prefabs.utils.DEUnderwaterProcessor;
 import com.legacy.structure_gel.util.ConfigTemplates;
 import com.legacy.structure_gel.worldgen.GelPlacementSettings;
 import com.legacy.structure_gel.worldgen.structure.GelTemplateStructurePiece;
@@ -11,12 +11,16 @@ import net.minecraft.nbt.CompoundNBT;
 import net.minecraft.util.Mirror;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.Rotation;
+import net.minecraft.util.SharedSeedRandom;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.ChunkPos;
 import net.minecraft.util.math.MutableBoundingBox;
 import net.minecraft.world.ISeedReader;
 import net.minecraft.world.IServerWorld;
+import net.minecraft.world.biome.Biome;
+import net.minecraft.world.biome.provider.BiomeProvider;
 import net.minecraft.world.gen.ChunkGenerator;
+import net.minecraft.world.gen.feature.NoFeatureConfig;
 import net.minecraft.world.gen.feature.structure.StructureManager;
 import net.minecraft.world.gen.feature.structure.StructurePiece;
 import net.minecraft.world.gen.feature.template.BlockIgnoreStructureProcessor;
@@ -32,6 +36,11 @@ public class DEUnderwaterStructure extends DEBaseStructure {
 
     public DEUnderwaterStructure(ConfigTemplates.StructureConfig config, boolean generateNear00, DEStructurePiece[] resources) {
         super(config, DETerrainAnalyzer.GenerationType.underwater, generateNear00, resources);
+    }
+
+    @Override
+    protected boolean checkLocation(ChunkGenerator chunkGen, BiomeProvider biomeProvider, long seed, SharedSeedRandom sharedSeedRand, int chunkPosX, int chunkPosZ, Biome biomeIn, ChunkPos chunkPos, NoFeatureConfig config) {
+        return DETerrainAnalyzer.isUnderwater(chunkPos, chunkGen, 16);
     }
 
     @Override
@@ -56,18 +65,19 @@ public class DEUnderwaterStructure extends DEBaseStructure {
         public PlacementSettings createPlacementSettings(TemplateManager templateManager) {
             BlockPos sizePos = Objects.requireNonNull(templateManager.get(this.name)).getSize();
             BlockPos centerPos = new BlockPos(sizePos.getX() / 2, 0, sizePos.getZ() / 2);
-            return new GelPlacementSettings().setMaintainWater(false).setRotation(rotation).setMirror(Mirror.NONE).setRotationPivot(centerPos);
+            return new GelPlacementSettings().setMaintainWater(true).setRotation(rotation).setMirror(Mirror.NONE).setRotationPivot(centerPos);
         }
 
         @Override
         public void addProcessors(TemplateManager templateManager, PlacementSettings placementSettings) {
+            placementSettings.clearProcessors();
             placementSettings.addProcessor(BlockIgnoreStructureProcessor.STRUCTURE_BLOCK);
-            placementSettings.addProcessor(DEUtil.Processors.Underwater);
+            placementSettings.addProcessor(DEUnderwaterProcessor.Instance);
         }
 
         @Override
-        public boolean postProcess(ISeedReader world, StructureManager structureManager, ChunkGenerator chunkGenerator, Random rand, MutableBoundingBox bounds, ChunkPos chunkPos, BlockPos pos) {
-            return super.postProcess(world, structureManager, chunkGenerator, rand, bounds, chunkPos, pos);
+        public boolean postProcess(ISeedReader world, StructureManager structureManager, ChunkGenerator chunkGenerator, Random rand, MutableBoundingBox box, ChunkPos chunkPos, BlockPos pos) {
+            return super.postProcess(world, structureManager, chunkGenerator, rand, box, chunkPos, pos);
         }
 
         @Override @ParametersAreNonnullByDefault